Projekat 2/EventInfoService/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
pymongo==4.15.3
Werkzeug==3.1.3
